% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_dose.R
\name{calculate_pdd}
\alias{calculate_pdd}
\title{Calculate Number of Prescribed Daily Doses Dispensed}
\usage{
calculate_pdd(df, drug, patient_id_col = "patient_id",
  drug_id_col = "drug_id", qty_disp_col = "qty_disp",
  qty_per_day_col = "qty_per_day")
}
\arguments{
\item{df}{a data frame containing prescribing records  to be analysed -
records must contain at least a paitent ID, drug ID, a quantity dispensed
and a quantity per day instruction}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{drug_id_col}{a string, the name of the column in \code{df} containing the
drug IDs}

\item{qty_disp_col}{a string, the name of the column in \code{df} containing the
quantity of drug dispensed}

\item{qty_per_day_col}{a string, the name of the column in \code{df} containing
the quantity of drug to be taken per day}
}
\value{
a modified version of \code{df} containing records for the drug of
  interest with the \code{dd_disp} field added
}
\description{
Calculates the number of Prescribed Daily Doses (PDDs) dispensed based on
individual patient instructions on how many tablets should be taken per day,
and the number of tablets dispensed, and adds this as a new column to the
data frame.
}
\examples{
\\code{}

}
