% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_dose.R
\name{dd_sum}
\alias{dd_sum}
\title{Calculate Cumulative Daily Dose}
\usage{
dd_sum(df, drug, threshold = 1, patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date",
  dd_disp_col = "dd_disp", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records  to be analysed -
records must contain at least a paitent ID, drug ID, a prescription date
and the number of daily doses disepnsed}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{threshold}{a number representing the minimum number of daily doses that
must be present for the patient to be considered exposed}

\item{patient_id_col}{a string, the name of the column in \code{df} containing
the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing the
drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df} containing
the prescption date}

\item{dd_disp_col}{a string, the name of the column in \code{df} containing the
number of daily doses dispensed}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing patient IDs and the corresponding number of
  prescriptions, number of daily doses dispensed and the date of the first
  prescription of the drug of interest
}
\description{
Calculates the cumulative number of daily doses of a drug of interest
dispensed for each patient. Results can be limited to only patients who have
been prescribed more than a desired threshold number of daily doses.
}
\examples{
\\code{}
\\code{}

}
