% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ever_use.R
\name{ever_use}
\alias{ever_use}
\title{Determine Ever Use of Drug(s) of Interest}
\usage{
ever_use(df, drug, summary = FALSE, threshold = 1,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a paitent ID, drug ID and a prescription date}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{summary}{logical, if TRUE the corresponding prescription counts and
first prescription dates will be returned alongside patient IDs}

\item{threshold}{a number representing the minimum number of prescriptions
that must be present for the patient to be considered exposed}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing patient IDs who match the selected criteria,
  and if \code{summary} is TRUE the corresponding prescription counts and
  dates
}
\description{
Determines which patients within a data frame have been exposed to the drug
of interest, based on having at least the desired number of prescriptions
matching the drug identifier of interest, with option to also provide counts
and dates of the first matching prescription.
}
\examples{
\\code{ever_use(synth_presc, drug = "CITALOPRAM", drug_id_col = "approved_name")}
\\code{ever_use(synth_presc, drug = "212000", summary = TRUE, threshold = 10,
drug_id_col = "bnf_paragraph", date_format = "\%d/\%m/\%Y")}

}
