% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{new_users_var}
\alias{new_users_var}
\title{Determine New vs. Ongoing Use of a Drug of Interest, Individual Start Date}
\usage{
new_users_var(df, df2, drug, timeframe = 0, return_all = TRUE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date_x", ev_date_col = "ev_date_1",
  date_format)
}
\arguments{
\item{df}{a data frame containing prescription records to be analysed -
records must contain at least a patient ID, drug ID and prescription date}

\item{df2}{a data frame containing dates to be used as patient specific start
dates - needs to contain at least patient IDs and event dates}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{timeframe}{a number representing the length, in days, of the timeframe
before the start date to check for previous prescriptions}

\item{return_all}{a logical, if TRUE the function will return data for all
patients as well as flags indicating if they were exposed during follow-up
and if they are new users. If FALSE, returns only new users exposed during
follow-up}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescription date}

\item{ev_date_col}{a string, the name of the column in \code{df2}
containing the start dates}

\item{date_format}{a string, the format of the dates contained in \code{df}
and  \code{df2}}
}
\value{
A data frame containing patient IDs, flags indicating if the patient
  was a new user at the start date and if they were exposed during follow-up,
  and where applicable the date of the first prescription during follow-up
}
\description{
Classifies patients as new or ongoing users of the drug(s) of interest based
on whether or not they have prescriptions within a set number of days of a
patient-specific follow-up start date
}
\examples{
new_users_var(synth_presc, df2 = synth_events,
drug = "SIMVASTATIN", timeframe = 90,
presc_date_col = "presc_date", drug_id_col = "approved_name",
ev_date_col = "start_date")
}
