% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{presc_by_time}
\alias{presc_by_time}
\title{Prescription Time Trends}
\usage{
presc_by_time(df, drug = "*", group = "Y", drug_id_col = "drug_id",
  presc_date_col = "presc_date", dd_disp_col = "dd_disp", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a drug identifier and a prescription date}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{group}{a string corresponding to the desired grouping for results -
either month ("M"), quarter ("Q"), semester ("S") or year ("Y")}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{dd_disp_col}{a string, the name of the column in \code{df} containing
the number of daily doses dispensed}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing the selected time intervals and number of
  prescriptions per interval
}
\description{
Function creates a breakdown of the number of prescriptions matching a chosen
drug identifier over time - either broken down by month, quarter, semester or
year.
}
\examples{
\\code{presc_by_time(synth_presc, drug = "SIMVASTATIN", drug_id_col = "approved_name", dd_disp_col = "ddd_dispensed", date_format = "\%d/\%m/\%Y")}
\\code{presc_by_time(synth_presc, drug = "212000", group = "M", drug_id_col = "bnf_paragraph", dd_disp_col = "ddd_dispensed", date_format = "\%d/\%m/\%Y")}

}
