% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{presc_per_patient}
\alias{presc_per_patient}
\title{Prescriptions Per Patient Summary Statistics}
\usage{
presc_per_patient(df, drug = "*", patient_id_col = "patient_id",
  drug_id_col = "drug_id")
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a drug identifier}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}
}
\value{
a dat aframe containing the mean, median, minimum and maximum number
  of prescriptions per patient
}
\description{
Calculates the mean, median, minimum and maximum numbers of prescriptions per
patient within the data - either for all drugs, or for those matching the
specified ID
}
\examples{
presc_per_patient(synth_presc, drug_id_col = "approved_name")
presc_per_patient(synth_presc, drug = "SIMVASTATIN", drug_id_col = "approved_name")

}
