% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{presc_top_drugs}
\alias{presc_top_drugs}
\title{Produce a List of the Most Commonly Prescribed Drugs}
\usage{
presc_top_drugs(df, rank = 10, drug_id_col = "drug_id")
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a drug identifier}

\item{rank}{the number of drugs to be included in the list}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}
}
\value{
a data frame containing a list of drug IDs and corresponding number
  of prescriptions
}
\description{
Function produces a list of the most commonly prescribed drugs and the number
of prescriptions for each, based on unique drug IDs. The number of results
returned can be adjusted.
}
\examples{
\\code{presc_top_drugs(synth_presc, drug_id_col = "approved_name")}
\\code{presc_top_drugs(synth_presc, rank = 5, drug_id_col = "bnf_item_code")}

}
