% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{refill_gap}
\alias{refill_gap}
\title{Determine Drug Persistence with Refill Gap Only}
\usage{
refill_gap(df, drug, gap, threshold = 0, summary = FALSE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date_x", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records  to be analysed -
records must contain at least a paitent ID, drug ID and a prescription date}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{gap}{a number, the length of the allowable gap in days between two
prescriptions before the patient is considered to have discontinued use of
the drug}

\item{threshold}{a number, the minimum length of period of exposure, in days,
to be included in the results}

\item{summary}{a logical, if true the function will return a summary of all
exposure periods instead of the individual periods}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing patient IDs and corresponding periods of
  exposure (consisting of first and last prescriptions, number of
  prescriptions and lengths of exposure) which match the chosen refill gap
  and minimum length
}
\description{
This function determines periods of persistent use of the drug of interest
based on the gaps between prescriptions. Gaps which exceed the selected
allowable gap length result in discontinuation points, and the function
returns patient ID's and periods of exposure that are over the selected
minimum length.
}
\examples{
refill_gap(synth_presc, drug = "SIMVASTATIN", gap = 30,
drug_id_col = "approved_name", presc_date_col = "presc_date",
date_format = "\%d/\%m/\%Y")
refill_gap(synth_presc, drug = "103050", gap = 30, threshold = 60, summary = TRUE,
drug_id_col = "bnf_paragraph", presc_date_col = "presc_date",
date_format = "\%d/\%m/\%Y")

}
