% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{refill_gap_dd}
\alias{refill_gap_dd}
\title{Determine Drug Persistence with Refill Gap and Coverage}
\usage{
refill_gap_dd(df, drug, gap, dd_factor = 1, threshold = 0,
  stockpile = FALSE, patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date",
  dd_disp_col = "dd_disp", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records  to be analysed -
records must contain at least a paitent ID, drug ID, a prescription date
and a number of daily doses dispensed}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{gap}{a number, the length of the allowable gap in days between two
prescriptions before the patient is considered to have discontinued use of
the drug}

\item{dd_factor}{a number, a multiplication factor applied to the dd_disp
field to calculate prescription duration - used, for example, if using an
assumption of 0.5 DDDs per day}

\item{threshold}{a number, the minimum length of period of exposure, in days,
to be included in the results}

\item{stockpile}{a logical, if TRUE the function will carry over leftover
medication from one prescription period to the next when determining
duration}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{dd_disp_col}{a string, the name of the column in \code{df} containing
the number of daily doses dispensed}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing patient IDs and corresponding periods of
  exposure (consisting of first and last prescriptions, number of
  prescriptions and lengths of exposure) which match the chosen refill gap
  and minimum length
}
\description{
This function determines periods of persistent use of the drug of interest
based on the gaps between prescriptions, taking into account the coverage of
those prescriptions. Prescription duration is defined based on the number of
daily doses dispensed, and there is the option to carry stockpiled
medications from one prescription over to the next . Gaps which exceed the
selected allowable gap length result in discontinuation points, and the
function returns patient ID's and periods of exposure that are over the
selected minimum length.
}
\examples{
\\code{refill_gap_dd(refill_gap_dd(synth_presc, drug = "OMEPRAZOLE", gap = 30, drug_id_col = "approved_name", dd_disp_col = "ddd_dispensed", date_format = "\%d/\%m/\%Y"))}
\\code{refill_gap_dd(synth_presc, drug = "CITALOPRAM", gap = 30, dd_factor = 1, threshold = 60, stockpile = TRUE, drug_id_col = "approved_name", dd_disp_col = "ddd_dispensed", date_format = "\%d/\%m/\%Y")}

}
