% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_fixed}
\alias{uat_fixed}
\title{Use at Time Point, Fixed Date Range}
\usage{
uat_fixed(df, drug, date_1, timeframe = 0, forward = TRUE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date_x", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a patient ID, drug ID and prescription date}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{date_1}{a string containing a date to base the follow-up window on}

\item{timeframe}{a number representing the length of the follow-up window. If
0 as by default any prescriptions before or after (depending on the value
of \code{forward}) the date entered in \code{date_1} will be considered}

\item{forward}{a logical, if TRUE the value of \code{timeframe} will be added
to the value of \code{date_1} to generate an end date for the follow-up
window. If FALSE, the value will be subtracted to generate a start date}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing a list of patient IDs for patients who have
  at least one prescription matching the criteria, the number of
  prescriptions within the defined date window and the date of the first
  matching prescription
}
\description{
Determines which patients have been exposed to the drug(s) of interest within
a timeframe of interest. This function applies the same timeframe to all
prescriptions - by default it considers all prescriptions before or after the
date provided, but a number of days can be entered to give a more specific
timeframe.
}
\examples{
uat_fixed(synth_presc, drug = "ATORVASTATIN", date_1 = "01/07/2020",
drug_id_col = "approved_name", presc_date_col = "presc_date", date_format = "\%Y-\%m-\%d")
uat_fixed(synth_presc, drug = "SIMVASTATIN", date_1 = "01/01/2021", timeframe = 180,
forward = FALSE, drug_id_col = "approved_name", presc_date_col = "presc_date",
date_format = "\%Y-\%m-\%d")

}
