% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_fixed_events}
\alias{uat_fixed_events}
\title{Use At Time Point, Fixed Range From Patient Event Date}
\usage{
uat_fixed_events(df, df2, drug, timeframe = 0, forward = TRUE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date_x", ev_date_col = "ev_date_1",
  date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a patient ID, drug ID, prescription date}

\item{df2}{a data frame containing event dates - should contain at least
patient IDs and an event date, can also include event codes}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{timeframe}{a number representing the length of the follow-up window. If
0 as by default any prescriptions before or after (depending on the value
of \code{forward}) the patient's event date, \code{presc_date}, will be
considered}

\item{forward}{a logical, if TRUE the value of \code{timeframe} will be added
to the value of \code{presc_date} to generate an end date for the follow-up
window. If FALSE, the value will be subtracted to generate a start date}

\item{patient_id_col}{a string, the name of the column in \code{df} and
\code{df2} containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the date the drug was prescribed}

\item{ev_date_col}{a string, the name of the column in \code{df2} containing
the event dates}

\item{date_format}{a string, the format of the dates contained in \code{df}
and \code{df2}}
}
\value{
data frame containing a list of patient IDs for patients who have at
  least one prescription matching the criteria, the number of prescriptions
  within the defined date window and the date of the first matching
  prescription
}
\description{
Determines which patients have been exposed to the drug(s) of interest within
a timeframe of interest. This function applies the same timeframe to
patient-specific event dates. By default it considers all prescriptions
before or after the event date, but a number of days can be entered to give a
more specific timeframe.
}
\examples{
uat_fixed_events(df = synth_presc, df2 = synth_events, drug = "SIMVASTATIN", forward = TRUE,
drug_id_col = "approved_name", presc_date_col = "presc_date", ev_date_col = "event_1",
date_format = "\%Y-\%m-\%d")

}
