% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_gap}
\alias{uat_gap}
\title{Two Prescriptions Within a Desired Timeframe}
\usage{
uat_gap(df, drug, timeframe, return_all = FALSE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date_x", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a patient ID, drug ID and prescription date}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{timeframe}{a number representing the desired maximum gap between
prescriptions}

\item{return_all}{logical, if TRUE function returns data for all patient IDs
with a flag indicating which patients did and did not meet the threshold
for exposure}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing the patient IDs and first two prescription
  dates for patients who meet the definition of exposure
}
\description{
Classifies patients as exposed or unexposed based on whether they received
two prescriptions for the drug of interest within a desired number of days of
each other
}
\examples{
uat_gap(df = synth_presc, drug = "SIMVASTATIN", timeframe = 14,
patient_id_col = "patient_id", drug_id_col = "approved_name",
presc_date_col = "presc_date", date_format = "\%Y-\%m-\%d")
}
