% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_new_users}
\alias{uat_new_users}
\title{Determine New vs. Ongoing Use of a Drug of Interest}
\usage{
uat_new_users(df, drug, start_date = NULL, timeframe = 0,
  return_all = TRUE, patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date_x",
  date_format)
}
\arguments{
\item{df}{a data frame containing prescription records to be analysed -
records must contain at least a patient ID, drug ID and prescription date}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{start_date}{a string containing the date to start follow-up - any
prescriptions on or after this date will be included. If NULL, all
prescriptions will be included}

\item{timeframe}{a number representing the length, in days, of the timeframe
before the patients' first prescription to check for previous prescriptions}

\item{return_all}{a logical, if TRUE the function will return all patients
with at least 1 prescription for the drug of interest, along with a flag
indicating whether or not the are new users}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescription date}

\item{date_format}{a string, the format of the dates contained in \code{df}
and \code{df2}}
}
\value{
A data frame containing patient IDs, the date of the first
  prescription and number of prescriptions during the follow up period, and
  if requested the flag specifying whether or not the patient is a new user
  of the drug of interest
}
\description{
Classifies patients as new or ongoing users of the drug(s) of interest based
on whether or not they have prescriptions within a user-defined lookback
period.
}
\examples{
uat_new_users(synth_presc, drug = "SIMVASTATIN", start_date = "01/01/2021", timeframe = 30,
presc_date_col = "presc_date", drug_id_col = "approved_name", date_format = "\%Y-\%m-\%d")
}
