% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_recent}
\alias{uat_recent}
\title{Define Recentness of Exposure at Event Date}
\usage{
uat_recent(df, df2, drug, timeframe, patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date_x",
  ev_date_col = "ev_date_1", date_format)
}
\arguments{
\item{df}{a data frame containing prescription records to be analysed;
records must contain at least a patient ID, drug ID and prescription date}

\item{df2}{a data frame containing medical event data to be analysed; records
must contain at least a patient ID and event date, maximum 1 event per
patient}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{timeframe}{a number, subtracted from the event date to define the start
of the rexcent use period}

\item{patient_id_col}{a string, the name of the column in \code{df} and
\code{df2} containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the date the drug was prescribed}

\item{ev_date_col}{a string, the name of the column in \code{df2} containing
the event dates}

\item{date_format}{a string, the format of the dates contained in \code{df}
and \code{df2}}
}
\value{
a data frame containing patient IDs, the number of prescriptions in
  each period and the use classification at the event date
}
\description{
Classifies patients as current or past users of the drug(s) of interest at an
event date based on a user-defined time window, and calculates the number of
prescriptions that fall within and outwith the recent use window.
}
\examples{
uat_recent(df = synth_presc, df2 = synth_events, drug = "OMEPRAZOLE", timeframe = 30,
drug_id_col = "approved_name", presc_date_col = "presc_date",
ev_date_col = "event_1", date_format = "\%Y-\%m-\%d")

}
