% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_dose.R
\name{dd_duration}
\alias{dd_duration}
\title{Calculate Prescription Durations}
\usage{
dd_duration(df, drug, dd_factor = 1, patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date",
  dd_disp_col = "dd_disp", date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records  to be analysed -
records must contain at least a paitent ID, drug ID, a prescription date}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{dd_factor}{a number, a multiplication factor applied to the dd_disp
field to calculate prescription duration}

\item{date_format}{a string, the format of the dates in \code{df}}

\item{patient_id}{a string, the name of the column in \code{df} containing
the patient IDs}

\item{drug_id}{a string, the name of the column in \code{df} containing the
drug IDs}

\item{presc_date}{a string, the name of the column in \code{df} containing
the prescption date}

\item{dd_disp}{a string, the name of the column in \code{df} containing the
number of daily doses dispensed}
}
\value{
a modified version of \code{df} with records for the drug of interest
  with the \code{duration} field added
}
\description{
Calculates the duration in days of individual prescriptions based on the
number of daily doses dispensed. A multiplication factor is applied to allow
for use in cases where, for example an assumption of less or more than 1 DDD
per day is required
}
\examples{
\\code{}
\\code{}

}
