% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{presc_data_summary}
\alias{presc_data_summary}
\title{Prescription and Patient Counts and Date Range}
\usage{
presc_data_summary(df, drug = "*", patient_id_col = "patient_id",
  drug_id_col = "drug_id", presc_date_col = "presc_date_x",
  date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a patient ID, a drug identifier and a
prescription date}

\item{drug}{a string containing a drug ID to be used to limit the prescribing
data to the drug(s) of interest, accepts regular expressions}

\item{patient_id_col}{a string, the name of the column in \code{df}
containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing the prescption date}

\item{date_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame containing the number of prescriptions, number of
  patients and the first and last prescription dates for the chosen data
  frame
}
\description{
Function produces a high-level summary of a data frame, containin the number
of patients and prescriptions in the data and the date range of the
prescriptions. Can either be used to obtain a summary of all prescriptions,
or limited to a drug or drugs of interest
}
\examples{
presc_data_summary(synth_presc, drug_id_col = "approved_name",
presc_date_col = "presc_date", date_format = "\%Y-\%m-\%d")
presc_data_summary(synth_presc, drug = "SIMVASTATIN", drug_id_col = "approved_name",
presc_date_col = "presc_date", date_format = "\%Y-\%m-\%d")

}
