% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_events.R
\docType{data}
\name{synth_events}
\alias{synth_events}
\title{Example medical event records for 100 patients}
\format{A data frame with 100 rows and 5 variavles}
\usage{
synth_events
}
\description{
\describe{ \item{patient_id}{a unique patient identifier}
  \item{start_date}{the start date of follow-up (the date of their first
  prescription in synth_presc)} \item{end_date}{the end date of follow-up}
  \item{event_1}{the date of the patient's first event, if applicable}
  \item{event_2}{the date of the patient's second event, if applicable}}
}
\details{
A dataset containing synthetic medical events records for the 100 patients
covered by the synth_presc dataset. Each row contains a patient ID, start and
end of follow-up dates and up to two event dates
}
\keyword{datasets}
