% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardising.R
\name{tidy_events}
\alias{tidy_events}
\title{Standardise Medical Event Dates}
\usage{
tidy_events(df, patient_id, ev_date_1, ev_code_1 = NULL,
  ev_date_2 = NULL, ev_code_2 = NULL, dates_format)
}
\arguments{
\item{df}{a data frame containing at least patient ID's and one corresponding
medical event date}

\item{patient_id}{a string, the name of the column in df containing the
patient IDs}

\item{ev_date_1}{a string, the name of the column containing the first event
dates}

\item{ev_code_1}{a string, the name of the column containing the medical code
for the first event}

\item{ev_date_2}{a string, the name of the column containing the second event
dates if present}

\item{ev_code_2}{a string, the name of the column containing the medical code
for the second event}

\item{dates_format}{a string, giving the format of the dates used in
\code{df}}
}
\value{
a data frame with standardised field names and formats
}
\description{
Adjusts the field names and data formats of medical event data to create a
data frame in a format suited for use with the use at time point functions in
this package. Called internally within the functions which make use of
medical event data, but can be used directly by the user to reduce the number
of arguments which need to be passed to other functions as these standardised
values are inluded as the default values in their code.
}
\examples{


}
