% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardising.R
\name{tidy_presc}
\alias{tidy_presc}
\title{Standardise Prescribing Data Field Names and Formats}
\usage{
tidy_presc(df, patient_id = NULL, drug_id = NULL, presc_date = NULL,
  dd_disp = NULL, qty_disp = NULL, qty_per_day = NULL, dates_format)
}
\arguments{
\item{df}{a data frame containing prescribing records}

\item{patient_id}{a string, the name of the column in \code{df} containing
the patient IDs}

\item{drug_id}{a string, the name of the column in \code{df} containing the
drug IDs}

\item{presc_date}{a string, the name of the column in \code{df} containing
the prescption date}

\item{dd_disp}{a string, the name of the column in \code{df} containing the
number of daily doses dispensed}

\item{qty_disp}{a string, the name of the column in \code{df} containing the
quantity of drug dispensed}

\item{qty_per_day}{a string, the name of the column in \code{df} containing
the quantity of drug to be taken per day}

\item{dates_format}{a string, the format of the dates in \code{df}}
}
\value{
a data frame with standardised field names and formats
}
\description{
Adjusts the field names and data formats of prescribing data to create a data
frame in a format suited for use with the other functions in this package -
mostly called internally by the other functions in this package, but can be
used directly  by the user to reduce the number of arguments which need to be
passed to other functions as these standardised values are inluded as the
default values in their code.
}
\examples{
\\code{tidy_synth <- tidy_presc(df = synth_presc, drug_id = "approved_name",
 dd_disp = "ddd_dispensed", qty_disp = "qty_dispensed",
 dates_format = "\%d/\%m/\%Y")}
\\code{tidy_synth <- tidy_presc(synth_presc, patient_id = "patient_id",
drug_id = "approved_name", presc_date = "presc_date",
dd_disp = "ddd_dispensed", qty_disp = "qty_dispensed",
qty_per_day = "qty_per_day", dates_format = "\%d/\%m/\%Y")}

}
