% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_at_time.R
\name{uat_var_events}
\alias{uat_var_events}
\title{Use at Time Point, Individual Patient Date Ranges}
\usage{
uat_var_events(df, df2, drug, forward = TRUE,
  patient_id_col = "patient_id", drug_id_col = "drug_id",
  presc_date_col = "presc_date", ev_date_1_col = "ev_date_1",
  ev_code_1_col = NULL, ev_date_2_col = "ev_date_2",
  ev_code_2_col = NULL, date_format)
}
\arguments{
\item{df}{a data frame containing prescribing records to be analysed -
records must contain at least a patient ID, drug ID, prescription date and
two event dates (\code{ev_date_1} and \code{ev_date_2})}

\item{drug}{a string corresponding to a drug identifier to be matched}

\item{forward}{a logical, if TRUE the follow-up window will start with
\code{ev_date_1} as the start date and \code{ev_date_2} as the end date. If
FALSE, the reverse will be true}

\item{patient_id_col}{a string, the name of the column in \code{df} and
\code{df2} containing the patient IDs}

\item{drug_id_col}{a string, the name of the column in \code{df} containing
the drug IDs}

\item{presc_date_col}{a string, the name of the column in \code{df}
containing}

\item{date_format}{a string, the format of the dates contained in \code{df}
and \code{df2}}

\item{ev_date_col_1}{a string, the name of the column in \code{df2}
containing the first event dates}

\item{ev_code_col_1}{a string, the name of the column in \code{df2}
containing the first event codes if present}

\item{ev_date_col_2}{a string, the name of the column in \code{df2}
containing the second event dates if present}

\item{ev_code_col_2}{a string, the name of the column in \code{df2}
containing the second event codes if present}
}
\value{
a data frame containing a list of patient IDs for patients who have
  at least one prescription matching the criteria, the number of
  prescriptions within the defined date window and the date of the first
  matching prescription
}
\description{
Determines which patients have been exposed to the drug(s) of interest within
a timeframe of interest. This function determines exposure by checking for
prescription dates between two patient-specific event dates.
}
\examples{
\\code{}

}
